Refactored sidebar click handlers into .js file. Updated user/index.js with useEffect dummy data.
