--- disco-db-demo/components/NoteContainer.jsx
@@ -1,28 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import Box from '@mui/material/Box';
-import Drawer from '@mui/material/Drawer';
-import CssBaseline from '@mui/material/CssBaseline';
-import AppBar from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
-import ListItem from '@mui/material/ListItem';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import InboxIcon from '@mui/icons-material/MoveToInbox';
-import MailIcon from '@mui/icons-material/Mail';
-import { Button } from '@mui/material';
-
-export default function NoteContainer (props){
-  console.log(props)
-  const num = Math.floor(Math.random() * 2)
-  return(
-    <Typography paragraph>
-      <h2>
-        {props.note[num].title}
-      </h2>
-      {props.note[num].content}
-    </Typography>
-  )
-};
