+++ disco-db-demo/pages/about.js
@@ -1,10 +1,7 @@
-import SideBar from "../components/sideBar.jsx"
-
 export default function About() {
   return (
     <div>
       <h1>About Page</h1>
-      <div><SideBar/></div>
       <h2>Hello</h2>
     </div>
   )
